+++ app/content/bookings/changelog.md
@@ -67,3 +67,4 @@
  - Update code to reflect changes to the F1::Http service
  - Update code to reflect changes to the F1:View service
  - Move `bookings.view.sql` to `app/` folder
+ - Fix bug where the nav calendar's date does not alays match the currently selected day's date.